audio_book/requirements.txt
django~=4.0
djangorestframework~=3.12.4
gunicorn==20.1.0
psycopg2-binary==2.9.1